Move publishSmartTransactionHook and publishBatchSmartTransactionHook into STX utils
